ecs: Add assertion when serial count is exhausted

+++ libs/ecs/src/entity_allocator_internal.h
@@ -6,7 +6,7 @@
 
 typedef struct {
   ThreadSpinLock lock;
-  u32            serialCounter;
+  u64            serialCounter;
   DynBitSet      freeIndices;
   usize          totalIndices;
 } EntityAllocator;
